add AppVersion class
